Add files via upload
In this Library Management System project, you can enter the record of new books and retrieve the details of books available in the library. You can issue the books to the students and  maintain their records. Late fine is charged for a student who returns the issued books after the  due date. Only one book is issued to students. New book is not issued to students those not  returned the last book.